User profile related ui updates
