+++ packages/client/src/App.tsx
@@ -51,7 +51,7 @@ import { ReportList } from '@client/views/SysAdmin/Performance/ReportList'
 import { WorkflowStatus } from '@client/views/SysAdmin/Performance/WorkflowStatus'
 import { TeamSearch } from '@client/views/SysAdmin/Team/TeamSearch'
 import { CreateNewUser } from '@client/views/SysAdmin/Team/user/userCreation/CreateNewUser'
-import { UserProfile } from '@client/views/SysAdmin/Team/user/UserProfile'
+import { UserProfile } from '@client/views/SysAdmin/Team/user/userProfilie/UserProfile'
 import { getTheme } from '@opencrvs/components/lib/theme'
 import ApolloClient from 'apollo-client'
 import { ConnectedRouter } from 'connected-react-router'
