+++ packages/client/src/user/queries.ts
@@ -52,6 +52,7 @@ export const GET_USER = gql`
       role
       type
       status
+      underInvestigation
       practitionerId
       primaryOffice {
         id
@@ -78,17 +79,22 @@ export const FETCH_TIME_LOGGED_METRICS_FOR_PRACTITIONER = gql`
     $timeEnd: String!
     $practitionerId: String!
     $locationId: String!
+    $count: Int!
   ) {
     fetchTimeLoggedMetricsByPractitioner(
       timeStart: $timeStart
       timeEnd: $timeEnd
       practitionerId: $practitionerId
       locationId: $locationId
+      count: $count
     ) {
-      status
-      trackingId
-      eventType
-      time
+      results {
+        status
+        trackingId
+        eventType
+        time
+      }
+      totalItems
     }
   }
 `
