+++ packages/components/src/components/interface/ListTable/ListTable.tsx
@@ -137,7 +137,7 @@ const ValueWrapper = styled.span<{
   display: flex;
   justify-content: ${({ alignment }) =>
     alignment === ColumnContentAlignment.RIGHT ? 'flex-end' : 'flex-start'};
-  align-items: center;
+  align-items: stretch;
   flex-shrink: 0;
   margin: auto 0;
   text-align: ${({ alignment }) => (alignment ? alignment.toString() : 'left')};
